# frontend/generated/openapi-current.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/credit-cards": {
      "get": {
        "tags": [
          "credit-cards-workspace"
        ],
        "summary": "Get Credit Cards",
        "description": "Get credit cards summary for the Credit Cards Intelligence Workspace.\n\nReturns aggregated data matching CreditCardsViewModel format.",
        "operationId": "get_credit_cards_api_v1_credit_cards_get",
        "parameters": [
          {
            "name": "statuses",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Statuses"
            }
          },
          {
            "name": "banks",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Banks"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Credit Cards Api V1 Credit Cards Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "credit-cards"
        ],
        "summary": "Create Card",
        "description": "Create a new credit card.",
        "operationId": "create_card_api_v1_credit_cards_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditCardCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Card Api V1 Credit Cards Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credit-cards/{card_id}": {
      "get": {
        "tags": [
          "credit-cards"
        ],
        "summary": "Get Card",
        "description": "Get credit card details.",
        "operationId": "get_card_api_v1_credit_cards__card_id__get",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Card Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditCardSummaryDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "credit-cards"
        ],
        "summary": "Update Card",
        "description": "Update credit card fields.",
        "operationId": "update_card_api_v1_credit_cards__card_id__put",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Card Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditCardUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Card Api V1 Credit Cards  Card Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "credit-cards"
        ],
        "summary": "Deactivate Card",
        "description": "Soft delete a credit card.",
        "operationId": "deactivate_card_api_v1_credit_cards__card_id__delete",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Card Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Deactivate Card Api V1 Credit Cards  Card Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credit-cards/{card_id}/statements": {
      "get": {
        "tags": [
          "credit-cards"
        ],
        "summary": "List Statements",
        "description": "Get statement history for a card.",
        "operationId": "list_statements_api_v1_credit_cards__card_id__statements_get",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Card Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 12,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatementDTO"
                  },
                  "title": "Response List Statements Api V1 Credit Cards  Card Id  Statements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "credit-cards"
        ],
        "summary": "Generate Statement",
        "description": "Generate a new statement for a credit card.",
        "operationId": "generate_statement_api_v1_credit_cards__card_id__statements_post",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Card Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatementGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatementDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credit-cards/{card_id}/outstanding": {
      "get": {
        "tags": [
          "credit-cards"
        ],
        "summary": "Get Outstanding",
        "description": "Get current outstanding balance.",
        "operationId": "get_outstanding_api_v1_credit_cards__card_id__outstanding_get",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Card Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Get Outstanding Api V1 Credit Cards  Card Id  Outstanding Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credit-cards/{card_id}/utilization": {
      "get": {
        "tags": [
          "credit-cards"
        ],
        "summary": "Get Utilization",
        "description": "Get credit utilization and available credit.",
        "operationId": "get_utilization_api_v1_credit_cards__card_id__utilization_get",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Card Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Get Utilization Api V1 Credit Cards  Card Id  Utilization Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/credit-cards/{card_id}/next-statement-date": {
      "get": {
        "tags": [
          "credit-cards"
        ],
        "summary": "Get Next Statement Date",
        "description": "Get the next expected statement date.",
        "operationId": "get_next_statement_date_api_v1_credit_cards__card_id__next_statement_date_get",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Card Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Get Next Statement Date Api V1 Credit Cards  Card Id  Next Statement Date Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credit-cards/{card_id}/payments": {
      "post": {
        "tags": [
          "credit-cards"
        ],
        "summary": "Record Payment",
        "description": "Record a payment on the latest open statement.",
        "operationId": "record_payment_api_v1_credit_cards__card_id__payments_post",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Card Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentRecordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatementDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/credit-cards/{card_id}/foreclosure": {
      "post": {
        "tags": [
          "credit-cards"
        ],
        "summary": "Quote Foreclosure",
        "description": "Quote foreclosure payoff for a credit card EMI.\n\nDelegates via credit_card_engine -> loan_engine.",
        "operationId": "quote_foreclosure_api_v1_credit_cards__card_id__foreclosure_post",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Card Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForeclosureRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForeclosureDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreditCardCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "bank": {
            "type": "string",
            "title": "Bank"
          },
          "card_last4": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{4}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Card Last4"
          },
          "credit_limit_paise": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Credit Limit Paise",
            "description": "Credit limit in paise"
          },
          "annual_fee_paise": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Annual Fee Paise",
            "default": 0
          },
          "interest_rate_bps": {
            "type": "integer",
            "maximum": 5000.0,
            "minimum": 0.0,
            "title": "Interest Rate Bps",
            "description": "Annual rate in basis points"
          },
          "billing_day": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 31.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Day"
          },
          "due_day_offset": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 1.0,
            "title": "Due Day Offset",
            "default": 21
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "account_id",
          "bank",
          "credit_limit_paise",
          "interest_rate_bps"
        ],
        "title": "CreditCardCreateRequest",
        "description": "Credit card creation request."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaymentRecordRequest": {
        "properties": {
          "payment_date": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "title": "Payment Date",
            "description": "ISO 8601 date string"
          },
          "amount_paise": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount Paise",
            "description": "Payment amount in paise"
          }
        },
        "type": "object",
        "required": [
          "payment_date",
          "amount_paise"
        ],
        "title": "PaymentRecordRequest",
        "description": "Request to record a payment on a statement."
      },
...
      "StatementGenerateRequest": {
        "properties": {
          "statement_date": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "title": "Statement Date",
            "description": "ISO 8601 date string for the statement"
          }
        },
        "type": "object",
        "required": [
          "statement_date"
        ],
        "title": "StatementGenerateRequest",
        "description": "Request to generate a new statement."
      },
...
    }
  }
}